JSDoc Type Checking
Added support for type checking in the editor by way of JSDoc type annotations! This is powered by TypeScript, and it's what larger projects out there use. It really helps with catching the unintended use of APIs different to how they were meant to be used, say like if you tried to call a variable that has a boolean, like it were a function. Rather than discovering that as an issue at runtime, it shows up beforehand as a type error in your editor.

The `tsconfig.json` is what allows the stronger analysis, and the JSDoc comments are what get validated to ensure the code is being called and used how it was annotated to be.

This is essentially just what TypeScript itself does, the general difference just being that the type annotations are part of the syntax of the language itself, rather than comments that are placed inside the plain JavaScript.

https://www.typescriptlang.org/docs/handbook/jsdoc-supported-types.html