JSZip npm Adjustments
npm packages should be managed with the `npm` terminal command, it has to use that because it allows the installation to have all of the dependencies it needs to run the script. Eventually we need to move from CDN urls over to a bundle step to smoothen things out for the production version of the site.

```sh
npm install jszip
```